requirements.txt
streamlit
openai
markdown
python-docx
openpyxl